requirements.txt
aiogram>=3.4
SQLAlchemy>=2.0
aiosqlite>=0.19
python-dotenv>=1.0
APScheduler>=3.10
